fix: address remaining CodeRabbit review findings
- Escape pipe characters in the eslint-plugin README rules table:
  the `numeric-tag-priority` row was rendering broken in GFM because
  the `'critical' | 'high' | 'low'` union sat inside an inline code
  span without escaping the table-cell delimiters.
- audit.vue's openFile no longer silently swallows /__open-in-editor
  failures; logs a warning to the console so a misconfigured Vite
  setup is debuggable.
- Consolidate the two CLI helper tests (validate-html.test.ts and
  validate-url.test.ts) into a single validate.test.ts. Both files
  exercised the shared validateHtml helper but were named after the
  commands, which was misleading. Added a docblock pointing at where
  the URL-fetch plumbing is actually exercised.
